agent-non-conversational/pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "MLflow-compatible agent server with FastAPI"
readme = "README.md"
authors = [
    { name = "Agent Developer", email = "developer@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "mlflow>=3.6.0rc0",
    "databricks-sdk>=0.63.0",
    "databricks-langchain>=0.7.0",
    "databricks-vectorsearch>=0.57.0",
    "databricks-agents>=1.4.0",
    "pydantic>=2.11",
    "python-dotenv",
]

[tool.uv]
prerelease = "allow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
